tests: extract logic to SchemaValidationTrait class
