+++ tests/Unit/API/REST/V3/Routes/Subscriptions/SubscriptionSchemaTest.php
@@ -15,6 +15,7 @@ use Give\Subscriptions\ValueObjects\SubscriptionStatus;
 use Give\Tests\RestApiTestCase;
 use Give\Tests\TestTraits\HasDefaultWordPressUsers;
 use Give\Tests\TestTraits\RefreshDatabase;
+use Give\Tests\Unit\API\REST\V3\SchemaValidationTrait;
 use WP_REST_Server;
 
 /**
@@ -24,6 +25,7 @@ class SubscriptionSchemaTest extends RestApiTestCase
 {
     use RefreshDatabase;
     use HasDefaultWordPressUsers;
+    use SchemaValidationTrait;
 
     /**
      * @unreleased
@@ -36,13 +38,13 @@ class SubscriptionSchemaTest extends RestApiTestCase
         $subscription = $this->createSubscription();
 
         // Get the schema via OPTIONS request
-        $schemaRoute = '/' . SubscriptionRoute::NAMESPACE . '/subscriptions';
+        $schemaRoute = '/' . SubscriptionRoute::NAMESPACE . '/' . SubscriptionRoute::BASE;
         $schemaRequest = $this->createRequest('OPTIONS', $schemaRoute, [], 'administrator');
         $schemaResponse = $this->dispatchRequest($schemaRequest);
         $schema = $schemaResponse->get_data();
 
         // Get actual subscription data
-        $dataRoute = '/' . SubscriptionRoute::NAMESPACE . '/subscriptions/' . $subscription->id;
+        $dataRoute = '/' . SubscriptionRoute::NAMESPACE . '/' . SubscriptionRoute::BASE . '/' . $subscription->id;
         $dataRequest = $this->createRequest(WP_REST_Server::READABLE, $dataRoute, [], 'administrator');
         $dataRequest->set_query_params(['includeSensitiveData' => true]);
         $dataResponse = $this->dispatchRequest($dataRequest);
@@ -69,13 +71,13 @@ class SubscriptionSchemaTest extends RestApiTestCase
         $subscription = $this->createSubscription();
 
         // Get the schema via OPTIONS request
-        $schemaRoute = '/' . SubscriptionRoute::NAMESPACE . '/subscriptions';
+        $schemaRoute = '/' . SubscriptionRoute::NAMESPACE . '/' . SubscriptionRoute::BASE;
         $schemaRequest = $this->createRequest('OPTIONS', $schemaRoute, [], 'administrator');
         $schemaResponse = $this->dispatchRequest($schemaRequest);
         $schema = $schemaResponse->get_data();
 
         // Get actual collection data
-        $dataRoute = '/' . SubscriptionRoute::NAMESPACE . '/subscriptions';
+        $dataRoute = '/' . SubscriptionRoute::NAMESPACE . '/' . SubscriptionRoute::BASE;
         $dataRequest = $this->createRequest(WP_REST_Server::READABLE, $dataRoute, [], 'administrator');
         $dataRequest->set_query_params(['includeSensitiveData' => true]);
         $dataResponse = $this->dispatchRequest($dataRequest);
@@ -104,7 +106,7 @@ class SubscriptionSchemaTest extends RestApiTestCase
         /** @var Subscription $subscription */
         $subscription = $this->createSubscription();
 
-        $route = '/' . SubscriptionRoute::NAMESPACE . '/subscriptions/' . $subscription->id;
+        $route = '/' . SubscriptionRoute::NAMESPACE . '/' . SubscriptionRoute::BASE . '/' . $subscription->id;
         $request = $this->createRequest(WP_REST_Server::READABLE, $route, [], 'administrator');
         $request->set_query_params(['includeSensitiveData' => true]);
         $response = $this->dispatchRequest($request);
@@ -118,174 +120,6 @@ class SubscriptionSchemaTest extends RestApiTestCase
         if (isset($data['renewsAt'])) {
             $this->validateDateFormat($data['renewsAt'], 'renewsAt');
         }
-    }
-
-    /**
-     * @unreleased
-     */
-    private function validateSchemaProperties($schema, $actualData)
-    {
-        if (!isset($schema['schema']['properties'])) {
-            $this->fail('Schema does not contain properties');
-        }
-
-        $schemaProperties = $schema['schema']['properties'];
-        $requiredProperties = $schema['schema']['required'] ?? [];
-
-        // Check that all required properties exist
-        foreach ($requiredProperties as $requiredProperty) {
-            $this->assertArrayHasKey(
-                $requiredProperty,
-                $actualData,
-                "Required property '{$requiredProperty}' is missing from response"
-            );
-        }
-
-        // Check that all schema properties that exist in actual data are properly defined
-        foreach ($actualData as $property => $value) {
-            if (isset($schemaProperties[$property])) {
-                // Property exists in schema, which is good
-                continue;
-            }
-
-            // Allow some dynamic properties that might not be in schema
-            $allowedDynamicProperties = ['_links', 'customFields'];
-            if (!in_array($property, $allowedDynamicProperties)) {
-                $this->fail("Property '{$property}' exists in response but not in schema");
-            }
-        }
-    }
-
-    /**
-     * @unreleased
-     */
-    private function validateDataTypes($schema, $actualData)
-    {
-        $schemaProperties = $schema['schema']['properties'];
-
-        foreach ($actualData as $property => $value) {
-            if (!isset($schemaProperties[$property])) {
-                continue; // Skip properties not in schema
-            }
-
-            // Skip properties that don't have a type (like objects with properties)
-            if (!isset($schemaProperties[$property]['type'])) {
-                continue;
-            }
-
-            $expectedType = $schemaProperties[$property]['type'];
-            $actualType = $this->getActualType($value);
-
-            // Handle oneOf schemas
-            if (isset($schemaProperties[$property]['oneOf'])) {
-                $this->validateOneOfSchema($schemaProperties[$property]['oneOf'], $value, $property);
-                continue;
-            }
-
-            // Handle array types
-            if (is_array($expectedType)) {
-                $this->assertContains(
-                    $actualType,
-                    $expectedType,
-                    "Property '{$property}' type '{$actualType}' should be one of: " . implode(', ', $expectedType)
-                );
-            } else {
-                $this->assertEquals(
-                    $expectedType,
-                    $actualType,
-                    "Property '{$property}' type mismatch. Expected: {$expectedType}, Actual: {$actualType}"
-                );
-            }
-        }
-    }
-
-    /**
-     * @unreleased
-     */
-    private function validateEnumValues($schema, $actualData)
-    {
-        $schemaProperties = $schema['schema']['properties'];
-
-        foreach ($actualData as $property => $value) {
-            if (!isset($schemaProperties[$property]['enum'])) {
-                continue;
-            }
-
-            $allowedValues = $schemaProperties[$property]['enum'];
-            $this->assertContains(
-                $value,
-                $allowedValues,
-                "Property '{$property}' value '{$value}' is not in allowed enum values: " . implode(', ', $allowedValues)
-            );
-        }
-    }
-
-    /**
-     * @unreleased
-     */
-    private function validateOneOfSchema($oneOfSchemas, $value, $propertyName)
-    {
-        $validTypes = [];
-        foreach ($oneOfSchemas as $schema) {
-            $validTypes[] = $schema['type'];
-        }
-
-        $actualType = $this->getActualType($value);
-        $this->assertContains(
-            $actualType,
-            $validTypes,
-            "Property '{$propertyName}' type '{$actualType}' should be one of: " . implode(', ', $validTypes)
-        );
-    }
-
-    /**
-     * @unreleased
-     */
-    private function validateDateFormat($dateValue, $propertyName)
-    {
-        if (is_string($dateValue)) {
-            // Should be RFC3339/ISO8601 format
-            $this->assertMatchesRegularExpression(
-                '/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?(?:Z|[+-]\d{2}:\d{2})$/',
-                $dateValue,
-                "Property '{$propertyName}' should be in RFC3339 format"
-            );
-        } elseif (is_null($dateValue)) {
-            // Null values are allowed for optional date fields
-            return;
-        } else {
-            $this->fail("Property '{$propertyName}' should be a string in RFC3339 format or null, got " . gettype($dateValue));
-        }
-    }
-
-    /**
-     * @unreleased
-     */
-    private function getActualType($value)
-    {
-        if (is_null($value)) {
-            return 'null';
-        }
-        if (is_bool($value)) {
-            return 'boolean';
-        }
-        if (is_int($value)) {
-            return 'integer';
-        }
-        if (is_float($value)) {
-            return 'number';
-        }
-        if (is_string($value)) {
-            return 'string';
-        }
-        if (is_array($value)) {
-            return 'array';
-        }
-        if (is_object($value)) {
-            return 'object';
-        }
-
-        return gettype($value);
     }
 
     /**
